# SharpTorch.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SharpTorch.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Torch/status": {
      "get": {
        "tags": [
          "Torch"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TorchStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorchStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorchStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/Torch/devices": {
      "get": {
        "tags": [
          "Torch"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Torch/models": {
      "get": {
        "tags": [
          "Torch"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TorchModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TorchModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TorchModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Torch/start-model": {
      "post": {
        "tags": [
          "Torch"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TorchModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TorchModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TorchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Torch/unload": {
      "delete": {
        "tags": [
          "Torch"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean",
                  "nullable": true
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "nullable": true
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean",
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/api/Torch/generate/stream": {
      "post": {
        "tags": [
          "Torch"
        ],
        "parameters": [
          {
            "name": "maxTokens",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 512
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TorchModel": {
        "type": "object",
        "properties": {
          "modelRootPath": {
            "type": "string",
            "nullable": true
          },
          "modelName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "modelPath": {
            "type": "string",
            "nullable": true
          },
          "jsonFiles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TorchStatus": {
        "type": "object",
        "properties": {
          "initialized": {
            "type": "boolean"
          },
          "cudaEnabled": {
            "type": "boolean"
          },
          "deviceId": {
            "type": "integer",
            "format": "int32"
          },
          "deviceName": {
            "type": "string",
            "nullable": true
          },
          "deviceType": {
            "type": "string",
            "nullable": true
          },
          "activeModel": {
            "$ref": "#/components/schemas/TorchModel"
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "Torch"
    }
  ]
}